move functionalities to methods crate lib
