+++ methods/Cargo.toml
@@ -18,6 +18,3 @@ clap = { version = "4.2.7", features = ["derive"] }
 hex = "0.4.3"
 risc0-build = { git = "https://github.com/risc0/risc0", branch = "release-0.15", features = ["guest-list"] }
 risc0-zkvm = { git = "https://github.com/risc0/risc0", branch = "release-0.15" }
-
-[patch.crates-io]
-radium = { git = "https://github.com/bitvecto-rs/radium", rev = "723bed5abd75994ee4b7221b8b12c9f4e77ce408" }
